chore: rename issue part 8 (#6952)
* chore: rename issue part 8

* chore: update

* chore: update

* chore: update
