+++ frontend/src/store/modules/index.ts
@@ -29,7 +29,6 @@ export * from "./idp";
 export * from "./user";
 export * from "./risk";
 export * from "./workspaceApprovalSetting";
-export * from "./review";
 export * from "./slowQuery";
 export * from "./role";
 export * from "./dbGroup";
