+++ frontend/src/store/modules/v1/index.ts
@@ -10,6 +10,7 @@ export * from "./sheet";
 export * from "./policy";
 export * from "./setting";
 export * from "./instance";
+export * from "./issue";
 export * from "./subscription";
 export * from "./actuator";
 export * from "./sql";
@@ -22,4 +23,3 @@ export * from "./activity";
 export * from "./bookmark";
 export * from "./inbox";
 export * from "./anomaly";
-export * from "./review";
